drafts: Extract drafts_overlay_ui module.
