+++ tools/test-js-with-node
@@ -90,6 +90,7 @@ EXEMPT_FILES = make_set(
         "web/src/desktop_integration.js",
         "web/src/dialog_widget.ts",
         "web/src/drafts.js",
+        "web/src/drafts_overlay_ui.js",
         "web/src/dropdown_widget.js",
         "web/src/echo.js",
         "web/src/emoji_picker.js",
